Mise en place de la securité JWT sur issue et comments

+++ monitoring/Issues/views.py
@@ -1,5 +1,6 @@
 from monitoring.Issues.models import Issue
 from monitoring.Issues.serializers import IssueSerializer
+from monitoring.permissions import IsContributor
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -7,8 +8,9 @@ from rest_framework.views import APIView
 
 class IssueProjectAPIView(APIView):
 
+    permission_classes = [IsContributor]
+
     def post(self, request, project_id):
-        user = self.request.user
         serializer = IssueSerializer(data=request.data, context={'request': request})
         if serializer.is_valid():
             serializer.save()
@@ -18,8 +20,8 @@ class IssueProjectAPIView(APIView):
     def get(self, request, project_id):
 
         try:
-            issues = Issue.objects.get(project=project_id)
-            serializer = IssueSerializer(issues)
+            issues = Issue.objects.filter(project=project_id)
+            serializer = IssueSerializer(issues, many=True)
             return Response(serializer.data, status=status.HTTP_200_OK)
         except Issue.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
@@ -27,10 +29,15 @@ class IssueProjectAPIView(APIView):
 
 class IssueUpdateDeleteAPIView(APIView):
 
+    permission_classes = [IsContributor]
+
     def put(self, request, project_id, issue_id):
 
         try:
             issue = Issue.objects.get(project=project_id, id=issue_id)
+            if issue.author_user != request.user:
+                return Response({"message": "Vous n'êtes pas l'auteur du problème."},
+                                status=status.HTTP_401_UNAUTHORIZED)
         except Issue.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
 
@@ -44,8 +51,11 @@ class IssueUpdateDeleteAPIView(APIView):
 
         try:
             issue = Issue.objects.get(project=project_id, id=issue_id)
+            if issue.author_user != request.user:
+                return Response({"message": "Vous n'êtes pas l'auteur du problème."},
+                                status=status.HTTP_401_UNAUTHORIZED)
             issue.delete()
-            return Response(content_type={"message": "Le commentaire a été supprimé avec succès."},
+            return Response({"message": "Le problème a été supprimé avec succès."},
                             status=status.HTTP_204_NO_CONTENT)
         except Issue.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
