+++ monitoring/comments/views.py
@@ -1,5 +1,7 @@
+from monitoring.Issues.models import Issue
 from monitoring.comments.models import Comment
 from monitoring.comments.serializers import CommentSerializer
+from monitoring.permissions import IsContributor
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -7,7 +9,13 @@ from rest_framework.views import APIView
 
 class CommentIssueAPIView(APIView):
 
+    permission_classes = [IsContributor]
+
     def post(self, request, project_id, issue_id):
+        issue = Issue.objects.get(id=issue_id)
+        if issue.assignee_user != request.user:
+            return Response({"message": "Vous n'êtes pas assigné pour créer un commentaire sur ce problème."},
+                            status=status.HTTP_401_UNAUTHORIZED)
         serializer = CommentSerializer(data=request.data, context={'request': request})
         if serializer.is_valid():
             serializer.save()
@@ -26,11 +34,13 @@ class CommentIssueAPIView(APIView):
 
 class CommentReadUpdateDeleteAPIView(APIView):
 
+    permission_classes = [IsContributor]
+
     def get(self, request, project_id, issue_id, comment_id):
 
         try:
-            comments = Comment.objects.get(issue=issue_id, comment=comment_id)
-            serializer = CommentSerializer(comments)
+            comment = Comment.objects.get(issue=issue_id, comment=comment_id)
+            serializer = CommentSerializer(comment)
             return Response(serializer.data, status=status.HTTP_200_OK)
         except Comment.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
@@ -39,6 +49,9 @@ class CommentReadUpdateDeleteAPIView(APIView):
 
         try:
             comment = Comment.objects.get(issue=issue_id, comment=comment_id)
+            if comment.author_user != request.user:
+                return Response({"message": "Vous n'êtes pas l'auteur du commentaire."},
+                                status=status.HTTP_401_UNAUTHORIZED)
         except comment.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
 
@@ -52,8 +65,11 @@ class CommentReadUpdateDeleteAPIView(APIView):
 
         try:
             comment = Comment.objects.get(issue=issue_id, comment=comment_id)
+            if comment.author_user != request.user:
+                return Response({"message": "Vous n'êtes pas l'auteur du commentaire."},
+                                status=status.HTTP_401_UNAUTHORIZED)
             comment.delete()
-            return Response(content_type={"message": "Le commentaire a été supprimé avec succès."},
+            return Response({"message": "Le commentaire a été supprimé avec succès."},
                             status=status.HTTP_204_NO_CONTENT)
         except Comment.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
